requirements.txt
django>=2.2.0,<2.3.0
# django>=3.0.0, <3.0.2
djangorestframework>=3.10.0,<=3.10.2
psycopg2>=2.8.0,<=2.8.4
phonenumbers>=8.11.0,<8.12.0
django-phonenumber-field>=3.9.0,<4.1.0
django-countries>=5.4.0,<=5.5.0
Pillow>=6.2.2,<=7.0.0

flake8>=3.7.0,<3.8.0